pyproject.toml
[tool.poetry]
name = "sql_mingle"
version = "0.1.0"
description = ""
authors = ["Brandon Braner <brandon.braner@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
typer = {extras = ["all"], version = "^0.4.0"}
rich = "^12.0.0"
python-dotenv = "^0.19.2"
pytest-dotenv = "^0.5.2"
psycopg = "^3.0.10"

[tool.poetry.dev-dependencies]
pytest = "^7.1"

[tool.pytest.ini_options]
env_files = [".env_test"]

[tool.poetry.scripts]
sqlmin = "sql_mingle.main:app"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
